Reorganize datastore files
