+++ addon/model/util/datastore/datastore.ts
@@ -1,10 +1,8 @@
 import * as RDF from '@rdfjs/types';
-import dataset, { FastDataset } from '@graphy/memory.dataset.fast';
 import ModelRange, {
   RangeContextStrategy,
 } from '@lblod/ember-rdfa-editor/model/model-range';
 import ModelNode from '@lblod/ember-rdfa-editor/model/model-node';
-import { NotImplementedError } from '@lblod/ember-rdfa-editor/utils/errors';
 import {
   ModelQuadObject,
   ModelQuadPredicate,
@@ -13,121 +11,19 @@ import {
   RdfaParser,
 } from '@lblod/ember-rdfa-editor/utils/rdfa-parser/rdfa-parser';
 import {
-  ConBlankNode,
   ConciseTerm,
   conciseToRdfjs,
-  ConLiteral,
-  ConNamedNode,
-  PrefixMapping,
   TermConverter,
 } from '@lblod/ember-rdfa-editor/model/util/concise-term-string';
 import { defaultPrefixes } from '@lblod/ember-rdfa-editor/config/rdfa';
-import { first, from, IterableX, single } from 'ix/iterable';
-import { filter, map, take } from 'ix/iterable/operators';
-
-export type TermSpec = SubjectSpec | PredicateSpec | ObjectSpec;
-export type SubjectSpec = RDF.Quad_Subject | ConNamedNode | ConBlankNode | null;
-export type PredicateSpec = RDF.Quad_Predicate | ConNamedNode | null;
-export type ObjectSpec =
-  | RDF.Quad_Object
-  | ConNamedNode
-  | ConBlankNode
-  | ConLiteral;
-type Primitive = number | string | boolean;
-
-function isPrimitive(thing: unknown): thing is Primitive {
-  return (
-    typeof thing === 'string' ||
-    typeof thing === 'boolean' ||
-    typeof thing === 'number'
-  );
-}
-
-export class ResultSet<I> implements Iterable<I> {
-  private engine: IterableX<I>;
-
-  constructor(iterable: Iterable<I>) {
-    this.engine = from(iterable);
-  }
-
-  first(): I | undefined {
-    return [...this.engine.pipe(take(1))][0];
-  }
-
-  map<T>(mappingFunc: (item: I) => T): ResultSet<T> {
-    return new ResultSet<T>(this.engine.pipe(map(mappingFunc)));
-  }
-
-  [Symbol.iterator](): Iterator<I> {
-    return this.engine[Symbol.iterator]();
-  }
-}
-
-export class TermMapping<T extends RDF.Term>
-  implements
-    Iterable<{
-      term: T;
-      nodes: ModelNode[];
-    }>
-{
-  private termMap: Map<T, ModelNode[]>;
-  private getPrefix: PrefixMapping;
-
-  constructor(map: Map<T, ModelNode[]>, getPrefix: PrefixMapping) {
-    this.termMap = map;
-    this.getPrefix = getPrefix;
-  }
-
-  /**
-   * Return the only mapping, if there is one.
-   * Throws if there is more than 1 mapping.
-   *
-   * Saves you from unpacking the iterator
-   * when you know for sure there can only be one answer,
-   * e.g. you matched on a subject and are requesting subjectNodes
-   */
-  single(): { term: T; nodes: ModelNode[] } | null {
-    return (
-      single(
-        from(this.termMap.entries()).pipe(
-          map((entry) => ({
-            term: entry[0],
-            nodes: entry[1],
-          }))
-        )
-      ) || null
-    );
-  }
-
-  [Symbol.iterator]() {
-    return from(this.termMap.entries())
-      .pipe(
-        map((entry) => ({
-          term: entry[0],
-          nodes: entry[1],
-        }))
-      )
-      [Symbol.iterator]();
-  }
-
-  /**
-   * Request the mapping for a specific term.
-   * @param term
-   */
-  get(term: TermSpec): ModelNode[] | null {
-    const convertedTerm = (
-      typeof term === 'string' ? conciseToRdfjs(term, this.getPrefix) : term
-    ) as T;
-    return (
-      first(
-        from(this.termMap.entries()).pipe(
-          filter((entry) => entry[0].equals(convertedTerm)),
-          map((entry) => entry[1])
-        )
-      ) || null
-    );
-  }
-}
+import { ResultSet } from '@lblod/ember-rdfa-editor/model/util/datastore/result-set';
+import { TermMapping } from '@lblod/ember-rdfa-editor/model/util/datastore/term-mapping';
+import {
+  isPrimitive,
+  ObjectSpec,
+  PredicateSpec,
+  SubjectSpec,
+} from '@lblod/ember-rdfa-editor/model/util/datastore/term-spec';
 
 export default interface Datastore {
   get dataset(): RDF.Dataset;
@@ -418,216 +314,4 @@ export class EditorStore implements Datastore {
   private getPrefix = (prefix: string): string | null => {
     return this._prefixMapping.get(prefix) || null;
   };
-}
-
-function isFastDataset(thing: unknown): thing is FastDataset {
-  // ts fails us here, see https://github.com/Microsoft/TypeScript/issues/21732
-  if (thing && 'isGraphyFastDataset' in (thing as Record<string, unknown>)) {
-    return (thing as FastDataset).isGraphyFastDataset;
-  }
-  return false;
-}
-
-function isGraphyDataset(thing: unknown): thing is GraphyDataset {
-  return thing instanceof GraphyDataset;
-}
-
-type QuadDataSet = RDF.Dataset<RDF.Quad, RDF.Quad>;
-
-/**
- * Spec-compliant wrapper for graphy
- */
-export class GraphyDataset implements QuadDataSet {
-  private _fastDataset: FastDataset;
-
-  constructor(quads?: RDF.Dataset | RDF.Quad[] | FastDataset | GraphyDataset) {
-    if (isFastDataset(quads)) {
-      this._fastDataset = quads;
-    } else if (isGraphyDataset(quads)) {
-      this._fastDataset = quads.fastDataset;
-    } else {
-      this._fastDataset = dataset();
-      if (quads) {
-        this._fastDataset.addAll(quads);
-      }
-    }
-  }
-
-  get size(): number {
-    return this.fastDataset.size;
-  }
-
-  get fastDataset(): FastDataset {
-    return this._fastDataset;
-  }
-
-  add(quad: RDF.Quad): this {
-    this.fastDataset.add(quad);
-    return this;
-  }
-
-  delete(quad: RDF.Quad): this {
-    this.fastDataset.delete(quad);
-    return this;
-  }
-
-  has(quad: RDF.Quad): boolean {
-    return this.fastDataset.has(quad);
-  }
-
-  match(
-    subject?: RDF.Quad_Subject | null,
-    predicate?: RDF.Quad_Predicate | null,
-    object?: RDF.Quad_Object | null,
-    graph?: RDF.Quad_Graph | null
-  ): QuadDataSet {
-    return new GraphyDataset(
-      this.fastDataset.match(subject, predicate, object, graph)
-    );
-  }
-
-  addAll(quads: QuadDataSet | RDF.Quad[]): this {
-    this._fastDataset.addAll(quads);
-    return this;
-  }
-
-  contains(other: QuadDataSet): boolean {
-    return this.intersection(other).size === other.size;
-  }
-
-  deleteMatches(
-    subject?: RDF.Quad_Subject,
-    predicate?: RDF.Quad_Predicate,
-    object?: RDF.Quad_Object,
-    graph?: RDF.Quad_Graph
-  ): this {
-    const matches = this.match(subject, predicate, object, graph);
-    const quads: RDF.Quad[] = [...matches];
-    this.fastDataset.deleteQuads(quads);
-    return this;
-  }
-
-  difference(other: QuadDataSet): QuadDataSet {
-    const gds = new GraphyDataset(other);
-    return new GraphyDataset(this.fastDataset.difference(gds.fastDataset));
-  }
-
-  equals(other: RDF.Dataset<RDF.Quad, RDF.Quad>): boolean {
-    const gds = new GraphyDataset(other);
-    return this.fastDataset.equals(gds.fastDataset);
-  }
-
-  every(
-    iteratee: (
-      quad: RDF.Quad,
-      dataset: RDF.Dataset<RDF.Quad, RDF.Quad>
-    ) => boolean
-  ): boolean {
-    for (const quad of this) {
-      if (!iteratee(quad, this)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  filter(
-    iteratee: (
-      quad: RDF.Quad,
-      dataset: RDF.Dataset<RDF.Quad, RDF.Quad>
-    ) => boolean
-  ): RDF.Dataset<RDF.Quad, RDF.Quad> {
-    const rslt = [];
-    for (const quad of this) {
-      if (iteratee(quad, this)) {
-        rslt.push(quad);
-      }
-    }
-    return new GraphyDataset(rslt);
-  }
-
-  forEach(
-    iteratee: (quad: RDF.Quad, dataset: RDF.Dataset<RDF.Quad, RDF.Quad>) => void
-  ): void {
-    for (const quad of this) {
-      iteratee(quad, this);
-    }
-  }
-
-  import(_stream: RDF.Stream<RDF.Quad>): Promise<this> {
-    throw new NotImplementedError();
-  }
-
-  intersection(other: QuadDataSet): QuadDataSet {
-    const gds = new GraphyDataset(other);
-    return new GraphyDataset(this.fastDataset.intersection(gds.fastDataset));
-  }
-
-  map(
-    iteratee: (
-      quad: RDF.Quad,
-      dataset: RDF.Dataset<RDF.Quad, RDF.Quad>
-    ) => RDF.Quad
-  ): RDF.Dataset<RDF.Quad, RDF.Quad> {
-    const result = [];
-    for (const quad of this) {
-      result.push(iteratee(quad, this));
-    }
-    return new GraphyDataset(result);
-  }
-
-  reduce<A>(
-    iteratee: (accumulator: A, quad: RDF.Quad, dataset: QuadDataSet) => A,
-    initialValue?: A
-  ): A {
-    const firstQuad = this[Symbol.iterator]().next().value as RDF.Quad | null;
-    let accumulator = initialValue || firstQuad;
-    // some bad typing in the spec causes these ugly casts
-    for (const quad of this) {
-      accumulator = iteratee(accumulator as A, quad, this);
-    }
-
-    return accumulator as A;
-  }
-
-  some(
-    iteratee: (
-      quad: RDF.Quad,
-      dataset: RDF.Dataset<RDF.Quad, RDF.Quad>
-    ) => boolean
-  ): boolean {
-    for (const quad of this) {
-      if (iteratee(quad, this)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  toArray(): RDF.Quad[] {
-    return [...this];
-  }
-
-  toCanonical(): string {
-    throw new NotImplementedError();
-  }
-
-  toStream(): RDF.Stream<RDF.Quad> {
-    throw new NotImplementedError();
-  }
-
-  toString(): string {
-    throw new NotImplementedError();
-  }
-
-  union(
-    quads: RDF.Dataset<RDF.Quad, RDF.Quad>
-  ): RDF.Dataset<RDF.Quad, RDF.Quad> {
-    const gds = new GraphyDataset(quads);
-    return new GraphyDataset(this.fastDataset.union(gds.fastDataset));
-  }
-
-  [Symbol.iterator](): Iterator<RDF.Quad> {
-    return this.fastDataset[Symbol.iterator]();
-  }
 }
